+++ expense-tracker-server/src/test/java/inzagher/expense/tracker/server/PersonServiceTests.java
@@ -34,7 +34,7 @@ public class PersonServiceTests {
     public void beforeEachTest() {
         bob = manager.storePerson("BOB");
         stan = manager.storePerson("STAN");
-        clothes = manager.storeCategory("CLOTHES", "CLOTHES PURCHASE", (byte)0, (byte)0, (byte)0);
+        clothes = manager.storeCategory("CLOTHES", "CLOTHES PURCHASE", 0, 0, 0);
         purchase = manager.storeExpense(2020, 10, 10, stan, clothes, 10.10D, "TEST CLOTHES PURCHASE");
         assertEquals(2L, manager.countPersons());
         assertEquals(1L, manager.countCategories());
